kube_sre_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-kube_sre_env"
version = "0.1.0"
description = "K8s SRE Self-Improving RL Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime
    "openenv-core[core]>=0.2.0",
    # K8s / GKE
    "kubernetes>=28.0.0",
    "google-auth>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "requests>=2.25.0",
    # LLM backends
    "huggingface_hub>=0.20.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "kube_sre_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["kube_sre_env", "kube_sre_env.server"]
package-dir = { "kube_sre_env" = ".", "kube_sre_env.server" = "server" }
